Made check for edit part ignore case
We encountered an issue where a hostname of "Sitemap.xml" broke
our site at runtime, since the system expected "sitemap.xml".

This fix makes that check ignore case. It also adds tests to prove the
fix. In order to make the method testable, it was made public
(this seemed like the least intrusive way of fixing it, and
it is a static method without side effects, so it being public
should not be an issue.